fix: update recommend link
